style: reduce duplication
